+++ scim2_models/rfc7643/enterprise_user.py
@@ -1,4 +1,5 @@
 from typing import Annotated
+from typing import ClassVar
 from typing import Literal
 from typing import Optional
 
@@ -26,7 +27,9 @@ class Manager(ComplexAttribute):
 
 
 class EnterpriseUser(Extension):
-    schemas: list[str] = ["urn:ietf:params:scim:schemas:extension:enterprise:2.0:User"]
+    scim_schema: ClassVar[str] = (
+        "urn:ietf:params:scim:schemas:extension:enterprise:2.0:User"
+    )
 
     employee_number: Optional[str] = None
     """Numeric or alphanumeric identifier assigned to a person, typically based