+++ scim2_models/rfc7643/group.py
@@ -1,4 +1,5 @@
 from typing import Annotated
+from typing import ClassVar
 from typing import Literal
 from typing import Optional
 from typing import Union
@@ -31,7 +32,7 @@ class GroupMember(MultiValuedComplexAttribute):
 
 
 class Group(Resource):
-    schemas: list[str] = ["urn:ietf:params:scim:schemas:core:2.0:Group"]
+    scim_schema: ClassVar[str] = "urn:ietf:params:scim:schemas:core:2.0:Group"
 
     display_name: Optional[str] = None
     """A human-readable name for the Group."""