+++ scim2_models/rfc7643/service_provider_config.py
@@ -1,5 +1,6 @@
 from enum import Enum
 from typing import Annotated
+from typing import ClassVar
 from typing import Optional
 
 from pydantic import Field
@@ -94,7 +95,9 @@ class AuthenticationScheme(ComplexAttribute):
 
 
 class ServiceProviderConfig(Resource):
-    schemas: list[str] = ["urn:ietf:params:scim:schemas:core:2.0:ServiceProviderConfig"]
+    scim_schema: ClassVar[str] = (
+        "urn:ietf:params:scim:schemas:core:2.0:ServiceProviderConfig"
+    )
 
     id: Annotated[
         Optional[str], Mutability.read_only, Returned.default, Uniqueness.global_