+++ scim2_models/rfc7644/error.py
@@ -1,4 +1,5 @@
 from typing import Annotated
+from typing import ClassVar
 from typing import Optional
 
 from pydantic import PlainSerializer
@@ -10,7 +11,7 @@ from .message import Message
 class Error(Message):
     """Representation of SCIM API errors."""
 
-    schemas: list[str] = ["urn:ietf:params:scim:api:messages:2.0:Error"]
+    scim_schema: ClassVar[str] = "urn:ietf:params:scim:api:messages:2.0:Error"
 
     status: Annotated[Optional[int], PlainSerializer(int_to_str)] = None
     """The HTTP status code (see Section 6 of [RFC7231]) expressed as a JSON