+++ tests/test_model_attributes.py
@@ -1,5 +1,6 @@
 import uuid
 from typing import Annotated
+from typing import ClassVar
 from typing import Optional
 
 import pytest
@@ -20,7 +21,7 @@ class Sub(ComplexAttribute):
 
 
 class Sup(Resource):
-    schemas: list[str] = ["urn:example:2.0:Sup"]
+    scim_schema: ClassVar[str] = "urn:example:2.0:Sup"
     dummy: str
     sub: Sub
     subs: list[Sub]
@@ -44,7 +45,7 @@ class Baz(ComplexAttribute):
 
 
 class Foo(Resource):
-    schemas: list[str] = ["urn:example:2.0:Foo"]
+    scim_schema: ClassVar[str] = "urn:example:2.0:Foo"
     sub: Annotated[ReturnedModel, Returned.default]
     bar: str
     snake_case: str
@@ -52,7 +53,7 @@ class Foo(Resource):
 
 
 class Bar(Resource):
-    schemas: list[str] = ["urn:example:2.0:Bar"]
+    scim_schema: ClassVar[str] = "urn:example:2.0:Bar"
     sub: Annotated[ReturnedModel, Returned.default]
     bar: str
     snake_case: str
@@ -60,7 +61,7 @@ class Bar(Resource):
 
 
 class Extension(Resource):
-    schemas: list[str] = ["urn:example:2.0:Extension"]
+    scim_schema: ClassVar[str] = "urn:example:2.0:Extension"
     baz: str